feat: non-blocking credential generation
