+++ vue.config.js
@@ -1,4 +1,5 @@
 module.exports = {
-  publicPath: process.env.NODE_ENV === "production" ? "/static/pes/" : "/",
-  outputDir: process.env.NODE_ENV === "production" ? "dist/pes" : "dist",
-};
+  publicPath: process.env.NODE_ENV === 'production' ? '/static/pes/' : '/',
+  outputDir: process.env.NODE_ENV === 'production' ? 'dist/pes' : 'dist',
+  configureWebpack: { output: { globalObject: 'this' } }
+}
